requirements.txt
Flask
pymongo
bcrypt
cloudinary
python-dotenv
gunicorn
Werkzeug
google-generativeai